rename config_utilities to config since util is implied by directory name
